Move `bip158` logic into `bip158` crate
There are some dependencies on `bitcoin` related traits that don't
exist in primitives, so I opted to move this crate along with `bitcoin`
as a dependency. Otherwise this is just a move of the module and test
vectors.
