+++ bip158/Cargo.toml
@@ -12,9 +12,20 @@ edition = "2021"
 rust-version = "1.74.0"
 exclude = ["tests", "contrib"]
 
+[features]
+default = ["std"]
+std = ["hashes/std", "internals/std", "io/std", "bitcoin/std"]
+
 [dependencies]
+bitcoin = { path = "../bitcoin/", default-features = false }
+hashes = { package = "bitcoin_hashes", version = "0.18.0", path = "../hashes", default-features = false }
+internals = { package = "bitcoin-internals", path = "../internals", default-features = false }
+io = { package = "bitcoin-io", path = "../io", default-features = false }
 
 [dev-dependencies]
+hex = { package = "hex-conservative", version = "0.3.0", default-features = false }
+hex_lit = "0.1.1"
+serde_json = "1.0.68"
 
 [package.metadata.docs.rs]
 all-features = true
